+++ pkg/server/config/config.go
@@ -50,7 +50,8 @@ const (
 	DefaultListObjectsIteratorCacheMaxResults = 10000
 	DefaultListObjectsIteratorCacheTTL        = 10 * time.Second
 
-	DefaultListObjectsOptimizationsEnabled = false
+	DefaultListObjectsOptimizationsEnabled            = false
+	DefaultListObjectsIntersectionAndExclusionEnabled = false
 
 	DefaultCacheControllerConfigEnabled = false
 	DefaultCacheControllerConfigTTL     = 10 * time.Second